Add basic support for adding and removing candidates.

+++ app/Http/Controllers/ElectionController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 use App\Http\Requests\StoreElectionRequest;
 use App\Http\Requests\UpdateElectionRequest;
@@ -49,7 +50,7 @@ class ElectionController extends Controller
         $election->end_at = $request->end_at;
         $request->user()->own_elections()->save($election);
 
-        return redirect(route('dashboard'));
+        return redirect(route('elections.edit', ['election' => $election]));
     }
 
     /**
@@ -71,7 +72,11 @@ class ElectionController extends Controller
      */
     public function edit(Election $election)
     {
-        //
+        if (! Gate::allows('edit-election', $election)) {
+            abort(403);
+        }
+
+        return view('elections.edit', ['election' => $election]);
     }
 
     /**
@@ -81,9 +86,23 @@ class ElectionController extends Controller
      * @param  \App\Models\Election  $election
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateElectionRequest $request, Election $election)
+    /* public function update(UpdateElectionRequest $request, Election $election) */
+    public function update(Request $request, Election $election)
     {
-        //
+        if (! Gate::allows('edit-election', $election)) {
+            abort(403);
+        }
+
+        $election->name = $request->name;
+        $election->description = $request->description;
+        $election->system = $request->system;
+        $election->public = $request->has('public');
+        $election->anonymous = $request->has('anonymous');
+        $election->start_at = $request->start_at;
+        $election->end_at = $request->end_at;
+        $request->user()->own_elections()->save($election);
+
+        return redirect(route('dashboard'));
     }
 
     /**
@@ -94,6 +113,10 @@ class ElectionController extends Controller
      */
     public function destroy(Election $election)
     {
+        if (! Gate::allows('edit-election', $election)) {
+            abort(403);
+        }
+
         $election->delete();
         return redirect(route('dashboard'));
     }
